// frontend/src/api/generated/zod/professores/professores.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SIFES API
 * API do Sistema do Instituto Federal de Esperantina — Diário de Aula
 * OpenAPI spec version: 1.0
 */
import * as zod from 'zod';





export const ProfessoresControllerCreateBody = zod.object({
  "nome": zod.string().min(1),
  "email": zod.email().describe('Também usado como login')
})

export const ProfessoresControllerCreateResponse = zod.object({
  "id": zod.string(),
  "userId": zod.string(),
  "nome": zod.string(),
  "email": zod.string(),
  "fotoUrl": zod.string().nullable(),
  "senhaInicial": zod.string().describe('Senha inicial gerada — mostrar ao admin uma única vez')
})

export const ProfessoresControllerFindAllResponseItem = zod.object({
  "id": zod.string(),
  "userId": zod.string(),
  "nome": zod.string(),
  "email": zod.string(),
  "fotoUrl": zod.string().nullable()
})
export const ProfessoresControllerFindAllResponse = zod.array(ProfessoresControllerFindAllResponseItem)

export const ProfessoresControllerFindOneParams = zod.object({
  "id": zod.string()
})

export const ProfessoresControllerFindOneResponse = zod.unknown()

export const ProfessoresControllerUpdateParams = zod.object({
  "id": zod.string()
})




export const ProfessoresControllerUpdateBody = zod.object({
  "nome": zod.string().min(1).optional(),
  "email": zod.email().optional().describe('Também usado como login')
})

export const ProfessoresControllerUpdateResponse = zod.unknown()

export const ProfessoresControllerRemoveParams = zod.object({
  "id": zod.string()
})

export const ProfessoresControllerRemoveResponse = zod.unknown()