// frontend/src/api/generated/models/professorDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SIFES API
 * API do Sistema do Instituto Federal de Esperantina — Diário de Aula
 * OpenAPI spec version: 1.0
 */

export interface ProfessorDto {
  id: string;
  userId: string;
  nome: string;
  email: string;
  /** @nullable */
  fotoUrl: string | null;
}